+++ truck-shapeops/src/fillet/mod.rs
@@ -1045,7 +1045,7 @@ fn create_simple_fillet() {
     )
     .into();
 
-    let v = Vertex::news(&[
+    let v = Vertex::news([
         Point3::new(-1.0, 0.0, 0.0),
         Point3::new(1.0, 0.0, 0.0),
         Point3::new(1.0, 1.0, 1.0),
@@ -1102,7 +1102,7 @@ fn create_fillet_with_side() {
         Point3::new(1.0, 1.0, 0.0),
         Point3::new(0.0, 1.0, 0.0),
     ];
-    let v = Vertex::news(&p);
+    let v = Vertex::news(p);
 
     let line = |i: usize, j: usize| {
         let bsp = BSplineCurve::new(KnotVec::bezier_knot(1), vec![p[i], p[j]]);
@@ -1180,7 +1180,7 @@ fn fillet_to_nurbs() {
         Point3::new(1.0, 0.0, 0.0),
         Point3::new(0.0, 1.0, 0.0),
     ];
-    let v = Vertex::news(&p);
+    let v = Vertex::news(p);
 
     let line = |i: usize, j: usize| {
         let bsp = BSplineCurve::new(KnotVec::bezier_knot(1), vec![p[i], p[j]]);
@@ -1287,7 +1287,7 @@ fn fillet_semi_cube() {
         Point3::new(1.1, 1.1, 0.0),
         Point3::new(0.0, 1.1, 0.0),
     ];
-    let v = Vertex::news(&p);
+    let v = Vertex::news(p);
 
     let line = |i: usize, j: usize| {
         let bsp = BSplineCurve::new(KnotVec::bezier_knot(1), vec![p[i], p[j]]);