+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The version is of the bottom crate `truck-rendimpl`.
 ## Unreleased
 
 - Review of the specifications for `IntersectionCurve`.
+- Fix STEP header description.
+- Fix some typos.
 - Implement `BSplineCurve::interpole`.
 - Implement `search_intersection_parameter` between surface and curve.
 - Add macros: `wire` and `shell`.
@@ -19,7 +21,7 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ### Latest `cargo upgrade`
 
-2024-12-01
+2024-12-06
 
 ## v0.6
 
@@ -38,7 +40,7 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ### Updated APIs
 
-- Add "periodic" identifer to `ParametricCurve` and `ParametricSurface`.
+- Add "periodic" identifier to `ParametricCurve` and `ParametricSurface`.
 - Remove the `Invertible` constraint from tessellating traits.
 - Features has been set up to use each module in `truck-meshalgo` separately.
 - Non-bounded parameter ranges has been supported. Updates `ParametricXXX` and `BoundedXXX`.